Fix : 퍼블릭 팀 distance 추가 -김성수 2025-02-15 10:52

+++ BE/PrePay/src/main/java/com/d111/PrePay/service/StoreService.java
@@ -62,7 +62,7 @@ public class StoreService {
     }
 
     // 경도 위도로 2km 이내이면 true
-    public double calDistance(float storeLongitude, float storeLatitude, float userLongitude, float userLatitude) {
+    public float calDistance(float storeLongitude, float storeLatitude, float userLongitude, float userLatitude) {
         final int EARTH_RADIUS_KM = 6371; // 지구 반지름 (km)
 
         // 라디안으로 변환
@@ -82,6 +82,6 @@ public class StoreService {
         double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
 
         double distanceKm = EARTH_RADIUS_KM * c; // 최종 거리 (km)
-        return distanceKm; // 2km 이내인지 확인
+        return (float) distanceKm; // 2km 이내인지 확인
     }
 }
